services/analyst-service/pyproject.toml
[project]
name = "llm-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Kolade Fajimi",email = "fajimikolade23@gmail.com"}
]
requires-python = ">=3.11,<4.0"
dependencies = [
    "grpcio (>=1.78.0,<2.0.0)",
    "llama-index-core (>=0.14.21,<0.15.0)",
    "llama-index-llms-ollama (>=0.10.1,<0.11.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic (>=2.13.3,<3.0.0)",
    "pydantic-settings (>=2.14.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aioboto3 (>=13.4.0,<14.0.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "protobuf (>=7.34.0,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "safety"
url = "https://pkgs.safetycli.com/repository/none-19107/pypi/simple/"
priority = "primary"

